Add backend flask skeleton
