+++ frontend/index.html
@@ -5,7 +5,7 @@
     <meta charset="UTF-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
     <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <link rel="stylesheet" href="style.css">
+    <link rel="stylesheet" href="static/style.css">
     <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.2.0/css/all.min.css"
         integrity="sha512-xh6O/CkQoPOWDdYTDqeRdPCVd1SpvCA9XXcUnZS2FmJNp1coAFzvtCN9BmamE+4aHK8yyUHUSCcJHgXloTyT2A=="
         crossorigin="anonymous" referrerpolicy="no-referrer" />
@@ -16,7 +16,7 @@
 <body id="backdrop-rotator">
     <nav class="navi">
         <div style="padding: 5px; width: 5%; float: left;">
-            <img id="logo" src="../images/logo.png">
+            <img id="logo" src="static/images/logo.png">
         </div>
         <div id = "mainLinks">
             <a href="#settings">Settings</a>  
@@ -29,7 +29,7 @@
     <p>Drag and drop the items into correct order.</p>
     <ul class="draggable-list" id="draggable-list"></ul>
     <button class="check-btn" id="check">Check</button>
-    <script src="script.js"></script>
+    <script src="./static/script.js"></script>
 </body>
 
 </html>
