+++ frontend/static/style.css
@@ -1,0 +1,140 @@
+@import url('https://fonts.googleapis.com/css?family=Lato&display=swap');
+
+:root {
+  --border-color: #e3e5e4;
+  --background-color: #c3c7ca;
+  --text-color: #34444f;
+}
+
+* {
+  box-sizing: border-box;
+}
+
+body {
+  background-color: #fff;
+  display: flex;
+  flex-direction: column;
+  align-items: center;
+  justify-content: flex-start;
+  height: 100vh;
+  margin: 0;
+  font-family: 'Lato', sans-serif;
+  
+  background-repeat: no-repeat;
+  background-position: center center;
+  background-attachment: fixed;
+  -webkit-background-size: cover;
+  -moz-background-size: cover;
+  -o-background-size: cover;
+  background-size: cover;
+}
+
+.draggable-list {
+  border: 1px solid var(--border-color);
+  color: var(--text-color);
+  padding: 0;
+  list-style-type: none;
+}
+
+.draggable-list li {
+  background-color: #fff;
+  display: flex;
+  flex: 1;
+}
+
+.draggable-list li:not(:last-of-type) {
+  border-bottom: 1px solid var(--border-color);
+}
+
+.draggable-list .number {
+  background-color: var(--background-color);
+  display: flex;
+  align-items: center;
+  justify-content: center;
+  font-size: 28px;
+  height: 80px;
+  width: 80px;
+}
+
+.draggable-list li.over .draggable {
+  background-color: #eaeaea;
+}
+
+.draggable-list .recipe-name {
+  margin: 0 20px 0 0;
+}
+
+.draggable-list li.right .recipe-name {
+  color: #3ae374;
+}
+
+.draggable-list li.wrong .recipe-name {
+  color: #ff3838;
+}
+
+.draggable {
+  cursor: pointer;
+  display: flex;
+  align-items: center;
+  justify-content: space-between;
+  padding: 15px;
+  flex: 1;
+}
+
+.check-btn {
+  background-color: var(--background-color);
+  border: none;
+  color: var(--text-color);
+  font-size: 16px;
+  padding: 10px 20px;
+  cursor: pointer;
+}
+
+.check-btn:active {
+  transform: scale(0.98);
+}
+
+.check-btn:focus {
+  outline: none;
+}
+
+
+/* Credits to W3 Schools for template: https://www.w3schools.com/howto/howto_js_topnav.asp */
+/* Add a black background color to the top navigation */
+.navi {
+  width: 100%;
+  height: 45px;
+  background-color: rgb(144, 100, 4);
+  overflow: hidden;
+  border-width: 5px;
+  border-bottom-width: 5px;
+  border-bottom: #132334;
+  border-bottom-style: solid;
+}
+
+/* Style the links inside the navigation bar */
+.navi a {
+  float: right;
+  color: #132334;
+  text-align: center;
+  padding: 14px 16px;
+  text-decoration: none;
+  font-size: 17px;
+}
+
+/* Change the color of links on hover */
+.navi a:hover {
+  background-color: rgb(181, 164, 89);
+  color: rgb(48, 16, 50);
+}
+
+/* Add a color to the active/current link */
+.navi a.active {
+  background-color: #67130b;
+  color: rgb(14, 55, 83);
+}
+
+#logo {
+  float: left;
+  height: 35px;
+}
